Seperated index.html into pages
