+++ brush-note.html
@@ -1,0 +1,50 @@
+<!doctype html>
+<html lang="he">
+  <head>
+    <meta charset="UTF-8" />
+    <title>PickMi - Anonymous Note Throwing</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
+  <body>
+    <!-- ====== BRUSH NOTE SCREEN ====== -->
+    <div id="brush-note-screen" class="container">
+      <img id="status-bar" src="./images/status-bar.png" />
+
+      <div class="app-bar">
+        <img
+          id="app-icon5"
+          src="/images/indicator-Day.png"
+          alt="App Icon"
+          class="App-icon"
+        />
+      </div>
+
+      <div class="write-card">
+        <div class="note-large note-brush">
+          <!-- כפתורי פעולה -->
+          <div class="button-group">
+            <button id="brush-back-btn" class="btn-link">← Back</button>
+
+            <button id="save-drawing-note-btn" class="btn-stamp">✉️</button>
+          </div>
+          <img src="/images/WriteBrush.png" alt="פתק לצביעה" />
+          <canvas id="note-canvas" width="420" height="300"></canvas>
+        </div>
+
+        <!-- אפשרויות מברשת -->
+        <div class="brush-controls">
+          <label for="brush-color"></label>
+          <input type="color" id="brush-color" value="#000000" />
+
+          <label for="brush-size"></label>
+          <input type="range" id="brush-size" min="1" max="30" value="5" />
+
+          <button id="clear-canvas-btn" title="נקה ציור">🗑️</button>
+        </div>
+      </div>
+      <img id="nav-bar" src="./images/navigation.png" alt="Navigation Bar" />
+    </div>
+    <script type="module" defer src="./js/brush-note.js"></script>
+  </body>
+</html>
