+++ forgot-password.html
@@ -1,0 +1,48 @@
+<!doctype html>
+<html lang="he">
+  <head>
+    <meta charset="UTF-8" />
+    <title>PickMi - Anonymous Note Throwing</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
+  <body>
+    <!-- ====== FORGOT PASSWORD ====== -->
+    <div class="container" id="forgot-password-screen">
+      <img id="status-bar" src="./images/status-bar.png" alt="סטטוס בר" />
+
+      <div class="app-bar">
+        <img
+          src="./images/Generic avatar.png"
+          alt="App Icon"
+          class="app-icon"
+        />
+      </div>
+
+      <div class="login-card">
+        <h1>Reset Password</h1>
+        <form id="forgot-form">
+          <div class="input-group">
+            <label for="forgot-email">Email </label>
+            <input
+              type="email"
+              id="forgot-email"
+              name="email"
+              class="auth-field"
+              required
+            />
+          </div>
+          <div id="forgot-error" class="error-msg hidden"></div>
+
+          <button type="submit" id="forgot-submit" class="btn-field">
+            Reset Password
+          </button>
+
+          <button id="forgot-back-to-login" class="btn-link">Back</button>
+        </form>
+      </div>
+      <img id="nav-bar" src="./images/navigation.png" alt="Navigation Bar" />
+    </div>
+    <script type="module" defer src="./js/forgot-password.js"></script>
+  </body>
+</html>
