+++ home.html
@@ -1,0 +1,58 @@
+<!doctype html>
+<html lang="he">
+  <head>
+    <meta charset="UTF-8" />
+    <title>PickMi - Anonymous Note Throwing</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
+  <body>
+    <!-- ====== HOME ====== -->
+
+    <div id="home-content" class="container">
+      <img id="status-bar" src="./images/status-bar.png" alt="סטטוס בר" />
+      
+
+      <div class="app-bar1">
+        <div class="button-wrapper logout">
+          <button id="logout-btn" class="btn-link" style="margin-top: 20px">
+            Logout
+          </button>
+        </div>
+
+        <div class="center-icon">
+          <img
+            src="/images/indicator-Day.png"
+            alt="App Icon"
+            class="App-icon"
+          />
+        </div>
+        <div class="button-wrapper map">
+          ,
+          <button id="to-map-btn" class="btn-link" style="margin-top: 20px">
+            Map
+          </button>
+        </div>
+      </div>
+
+      <div id="home-bg"></div>
+
+      <div id="notes-wrapper">
+        <div id="notes-list"></div>
+      </div>
+
+      <div id="new-note-btn" class="note-btn">
+        <img src="/images/WritePen (1).png" alt="פתח פתק חדש" />
+      </div>
+
+      <img id="nav-bar" src="./images/navigation.png" alt="Navigation Bar" />
+    </div>
+
+    <script
+      defer
+      src="https://maps.googleapis.com/maps/api/js?key=%VITE_GOOGLE_MAPS_API_KEY%&libraries=places,marker&callback=initHomeMap&loading=async"
+    ></script>
+
+    <script type="module" defer src="./js/home.js"></script>
+  </body>
+</html>
