+++ reset-sent.html
@@ -1,0 +1,35 @@
+<!doctype html>
+<html lang="he">
+  <head>
+    <meta charset="UTF-8" />
+    <title>PickMi - Anonymous Note Throwing</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
+  <body>
+    <!-- ====== RESET SENT CONFIRM ====== -->
+    <div class="container" id="reset-sent-screen">
+      <img id="status-bar" src="./images/status-bar.png" alt="סטטוס בר" />
+
+      <div class="app-bar">
+        <img
+          src="./images/Generic avatar.png"
+          alt="App Icon"
+          class="app-icon"
+        />
+      </div>
+
+      <div class="login-card">
+        <h1>Link Sent</h1>
+        <div class="success-msg" style="margin-bottom: 16px">
+          If there's an account associated with this email address, we've sent
+          you a link to reset your password.
+        </div>
+        <button id="sent-back" class="btn-link">Home</button>
+      </div>
+
+      <img id="nav-bar" src="./images/navigation.png" alt="Navigation Bar" />
+    </div>
+    <script type="module" defer src="./js/reset-sent.js"></script>
+  </body>
+</html>
