+++ signup.html
@@ -1,0 +1,57 @@
+<!doctype html>
+<html lang="he">
+  <head>
+    <meta charset="UTF-8" />
+    <title>PickMi - Anonymous Note Throwing</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
+  <body>
+    <!-- ====== SIGNUP ====== -->
+    <div class="container" id="signup-screen">
+      <img id="status-bar" src="./images/status-bar.png" alt="סטטוס בר" />
+
+      <div class="app-bar">
+        <img
+          src="./images/Generic avatar.png"
+          alt="App Icon"
+          class="app-icon"
+        />
+      </div>
+      <div class="login-card">
+        <h1>Sign Up</h1>
+        <form id="signup-form">
+          <div class="input-group">
+            <label for="signup-email">Email</label>
+            <input
+              type="email"
+              id="signup-email"
+              class="auth-field"
+              name="email"
+              required
+            />
+          </div>
+          <div class="input-group">
+            <label for="signup-password">Password</label>
+
+            <input
+              type="password"
+              id="signup-password"
+              class="auth-field"
+              name="password"
+              required
+            />
+          </div>
+          <div id="signup-error" class="btn-field"></div>
+          <button type="submit" class="btn-field">Save</button>
+          <button id="back-to-login" type="button" class="btn-link">
+            Back
+          </button>
+        </form>
+      </div>
+      <img id="nav-bar" src="./images/navigation.png" alt="Navigation Bar" />
+    </div>
+
+    <script type="module" defer src="./js/signup.js"></script>
+  </body>
+</html>
