+++ user-menu.html
@@ -1,0 +1,37 @@
+<!doctype html>
+<html lang="he">
+  <head>
+    <meta charset="UTF-8" />
+    <title>PickMi - Anonymous Note Throwing</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
+  <body>
+    <!-- ====== USER MENU SCREEN ====== -->
+    <div id="user-menu-screen" class="container">
+      <img id="status-bar" src="./images/status-bar.png" alt="סטטוס בר" />
+
+      <div class="app-bar">
+        <img
+          id="app-icon2"
+          src="/images/indicator-Day.png"
+          alt="App Icon"
+          class="App-icon"
+        />
+      </div>
+
+      <div class="settings-card">
+        <h2>Settings</h2>
+
+        <br />
+
+        <button id="delete-account-btn" class="btn-field">
+          🗑️ Delete Account
+        </button>
+        <button id="user-menu-back-btn" class="btn-link">home</button>
+      </div>
+      <img id="nav-bar" src="./images/navigation.png" alt="Navigation Bar" />
+    </div>
+    <script type="module" defer src="./js/user-menu.js"></script>
+  </body>
+</html>
