+++ wrong-password.html
@@ -1,0 +1,35 @@
+<!doctype html>
+<html lang="he">
+  <head>
+    <meta charset="UTF-8" />
+    <title>PickMi - Anonymous Note Throwing</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="./style/style.css" />
+  </head>
+  <body>
+    <!-- ====== WRONG PASSWORD ====== -->
+    <div class="container" id="wrong-password-screen">
+      <img id="status-bar" src="./images/status-bar.png" alt="סטטוס בר" />
+
+      <div class="app-bar">
+        <img
+          src="./images/Generic avatar.png"
+          alt="App Icon"
+          class="app-icon"
+        />
+      </div>
+      <div class="login-card">
+        <h1>Wrong Password</h1>
+        <div class="error-msg" style="margin-bottom: 20px">
+          Wrong password or email
+        </div>
+        <button id="wrong-back" class="btn-field">Home</button>
+        <button class="btn-link" id="wrong-forgot" class="btn-field">
+          Forgot Password
+        </button>
+      </div>
+      <img id="nav-bar" src="./images/navigation.png" alt="Navigation Bar" />
+    </div>
+    <script type="module" defer src="./js/wrong-password.js"></script>
+  </body>
+</html>
